Use std::vector instead of std::map for reading or writing an
arbitrary number of collections. The current implementation with std::map is not
easy to use. The collections get ordered alphabetically in std::map so the order
set in the steering file is lost, which means if there is some kind of
mapping (like one output collection for each input collection) then one has to
reproduce this map in the algorithm to make sure it's correct.