Get the available collections from IOSvc instead of peeking in the file
when running single-threaded to improve the performance since it's not necessary
to read the file before running to find how many events or which collections
exist. For running multithreaded, a fix has been discussed in Gaudi (because
there is not a way to tell the event loop to stop right now because there are no
more events), but will take some time.