Updates to INoiseCaloCellsTool and INoiseConstTool.
Make INoiseCaloCellsTool::addRandomCellNoise and filterCellNoise as well as
INoiseConstTool::getNoiseRMSPerCell and getNoiseRMSPerCell const.

Add new overloads for INoiseCaloCellsTool::addRandomCellNoise and
filterCellNoise that return a vector of cell/energy pairs rather than a set.

Impementations of this interface in k4RecCalorimeter have already been
adjusted appropriately:

https://github.com/HEP-FCC/k4RecCalorimeter/pull/177
https://github.com/HEP-FCC/k4RecCalorimeter/pull/183

Clients should not be affected.