Make the KeepDropSwitch understand type commands
Using "type" as second element of an output command will now make the
KeepDropSwitch take decisions on types as well. The isOn command gets a
second overload that takes a type name as well. Type names are matched
exactly without support for wildcards. We keep the old isOn overload
that only takes a single argument for easier testing without types and
for some backwards compatibility.