Add example that should fail but doesn't
The ExampleFailingSingleEvent test case should exit with a non-zero exit
code, because there is an exception that is thrown during the event
loop. However, since there is only a single event in the input file and
we use EvtMax=-1 this is changed into a user requested ScheduleStop and
a non-failing exit code.